fact-app/requirements.txt
flask>=3.0
psycopg2-binary>=2.9
confluent-kafka<=2.12
